1.4.0
added new lobby
turned falldamage off in the loading so people dont die to it
fixed the no bonus thing so it actually shows up

the split for 1.21.9+ was because they renamed one thing in the worldgen file, which blew up the whole datapack. thanks mojang. https://minecraft.wiki/w/Noise_router#History